+++ internal/templates/v1/models.go
@@ -7,7 +7,7 @@ import (
 )
 
 const (
-	appVersion = 49
+	appVersion = 50
 )
 
 const (
